Move category api endpoints under controllers
